requirements.txt
deta==1.1.0
pandas==1.5.3
plotly==5.12.0
python-dotenv==0.21.1
streamlit==1.17.0
streamlit_authenticator==0.1.5
streamlit_option_menu==0.3.2
